@@ -10,7 +10,7 @@ plugins {
 
 android {
   namespace = "com.trm.daylighter"
-  compileSdk = 35
+  compileSdk = 36
 
   defaultConfig {
     applicationId = "com.trm.daylighter"
