@@ -1,11 +1,12 @@
 package com.trm.daylighter
 
 import com.android.build.api.dsl.CommonExtension
-import java.io.File
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.dependencies
 import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.withType
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 /** Configure Compose-specific options */
 internal fun Project.configureAndroidCompose(commonExtension: CommonExtension<*, *, *, *, *, *>) {
@@ -14,7 +15,9 @@ internal fun Project.configureAndroidCompose(commonExtension: CommonExtension<*,
   commonExtension.apply {
     buildFeatures { compose = true }
 
-    kotlinOptions { freeCompilerArgs += buildComposeMetricsParameters() }
+    tasks.withType<KotlinCompile>().configureEach {
+      compilerOptions { freeCompilerArgs.addAll(buildComposeMetricsParameters()) }
+    }
 
     dependencies {
       val bom = libs.findLibrary("androidx-compose-bom").get()
@@ -26,10 +29,10 @@ internal fun Project.configureAndroidCompose(commonExtension: CommonExtension<*,
 
 private fun Project.buildComposeMetricsParameters(): List<String> {
   val metricParameters = mutableListOf<String>()
-  val enableMetricsProvider = project.providers.gradleProperty("enableComposeCompilerMetrics")
-  val enableMetrics = (enableMetricsProvider.orNull == "true")
+  val enableMetricsProvider = providers.gradleProperty("enableComposeCompilerMetrics")
+  val enableMetrics = enableMetricsProvider.orNull == "true"
   if (enableMetrics) {
-    val metricsFolder = File(project.buildDir, "compose-metrics")
+    val metricsFolder = layout.buildDirectory.dir("compose-metrics").get().asFile
     metricParameters.add("-P")
     metricParameters.add(
       "plugin:androidx.compose.compiler.plugins.kotlin:metricsDestination=" +
@@ -37,15 +40,16 @@ private fun Project.buildComposeMetricsParameters(): List<String> {
     )
   }
 
-  val enableReportsProvider = project.providers.gradleProperty("enableComposeCompilerReports")
-  val enableReports = (enableReportsProvider.orNull == "true")
+  val enableReportsProvider = providers.gradleProperty("enableComposeCompilerReports")
+  val enableReports = enableReportsProvider.orNull == "true"
   if (enableReports) {
-    val reportsFolder = File(project.buildDir, "compose-reports")
+    val reportsFolder = layout.buildDirectory.dir("compose-reports").get().asFile
     metricParameters.add("-P")
     metricParameters.add(
       "plugin:androidx.compose.compiler.plugins.kotlin:reportsDestination=" +
         reportsFolder.absolutePath
     )
   }
+
   return metricParameters.toList()
 }
