@@ -18,7 +18,7 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
 
       extensions.configure<ApplicationExtension> {
         configureKotlinAndroid(this)
-        defaultConfig.targetSdk = 35
+        defaultConfig.targetSdk = 36
         configureFlavors(this)
         configureGradleManagedDevices(this)
       }
