@@ -22,7 +22,7 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
 
       extensions.configure<LibraryExtension> {
         configureKotlinAndroid(this)
-        defaultConfig.targetSdk = 35
+        defaultConfig.targetSdk = 36
         configureFlavors(this)
         configureGradleManagedDevices(this)
       }
