@@ -15,7 +15,7 @@ class AndroidTestConventionPlugin : Plugin<Project> {
 
       extensions.configure<TestExtension> {
         configureKotlinAndroid(this)
-        defaultConfig.targetSdk = 35
+        defaultConfig.targetSdk = 36
         configureGradleManagedDevices(this)
       }
     }
